Refactor loadpoint.go
Put functions for pv mode in separate file
